% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiogramslice.R
\name{audiogramslice}
\alias{audiogramslice}
\title{Get high and low frequency hearing limits, best sensitivity,
and best frqeuency from an audiogram for a given
SPL level.}
\usage{
audiogramslice(cutoff = 35, Hz, threshold, plot = T)
}
\arguments{
\item{Hz}{numerical vector with frequencies of the audiogram}

\item{threshold}{numerical vector with the thresholds}

\item{plot}{logical term to decide whether to plot audiogram with cutoff values}
}
\description{
Get high and low frequency hearing limits, best sensitivity,
and best frqeuency from an audiogram for a given
SPL level.
}
\examples{
audiogramslice(Hz = Hz_, threshold = threshold_, plot = F)
lowlimit highlimit bestsensitivity   besthz
349.5812  752.2525        30.00561 499.8317
}
