% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert}
\usage{
convert(from, to, inputval)
}
\arguments{
\item{v1}{speed of first one}

\item{v2}{speed of second one}

\item{direction}{specify the direction as towards or away from each other}

\item{f}{frequency}
}
\description{
Convert between dB and Pa
}
\details{
S2 (time shifted sound), Tadd (time added), Time (time)
}
\examples{
timeshiftpressure(Amplitude = 1, freq = 2, distance = 340, c = 340)
1.002374
}
