% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBtoPa.R
\name{dBtoPa}
\alias{dBtoPa}
\title{Convert dB to Pascals}
\usage{
dBtoPa(dB, ref = 20)
}
\arguments{
\item{dB}{dB value}

\item{ref}{reference pressure in micropasals}
}
\description{
This function allows you to convert decibels to Pascals value. Defaults to reference pressure of 20 microPascals.
}
\examples{
dBtoPa(94,20)
1.002374
}
