% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hztocents.R
\name{hztocents}
\alias{hztocents}
\title{Convert a frequency interval to cents, semitones, and octaves}
\usage{
hztocents(f1, f2)
}
\arguments{
\item{f1}{second frequency}
}
\description{
Convert a frequency interval to cents, semitones, and octaves
}
\examples{
> hztocents(100,200)
[[1]]
[1] "1200 cents"

[[2]]
[1] "12 semitones"

[[3]]
[1] "1 ocataves"

}
