% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler equation pmotion.R, R/intensity ohms
%   law.R
\name{intensity}
\alias{intensity}
\title{Convert dB to Pascals}
\usage{
intensity(Pa = 1, Z = 340)

intensity(Pa = 1, Z = 340)
}
\arguments{
\item{dB}{decibel value}

\item{ref}{reference pressure in micropasals}
}
\description{
This function allows you to convert decibels to Pascals. Defaults to a reference pressure of 20 microPascals.

This function allows you to convert decibels to Pascals. Defaults to a reference pressure of 20 microPascals.
}
\examples{
dBtoPa(94,20)
1.002374
dBtoPa(94,20)
1.002374
}
