% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_array_Onset time diff.R
\name{onsettimediff}
\alias{onsettimediff}
\title{Calculates the sine wave and time shifted time wave. Output is a list: S1 (first sound),
S2 (time shifted sound), Tadd (time added), Time (time). Assumes angle of 0 between
sound source and two receivers.}
\usage{
onsettimediff(amplitude = 1, freq = 2, distance = 20, c = 340, plot = TRUE)
}
\arguments{
\item{c}{speed of sound}

\item{d}{distance between two ears or receivers}

\item{f}{frequency}
}
\description{
Calculates the sine wave and time shifted time wave. Output is a list: S1 (first sound),
S2 (time shifted sound), Tadd (time added), Time (time). Assumes angle of 0 between
sound source and two receivers.
}
\examples{
timeshiftpressure(Amplitude = 1, freq = 2, distance = 340, c = 340)
1.002374
}
