% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatodB.R
\name{PatodB}
\alias{PatodB}
\title{Convert Pa to dB}
\usage{
PatodB(Pa, ref = 20)
}
\arguments{
\item{Pa}{Pascals value}

\item{ref}{reference pressure in micropasals.}
}
\description{
Converts Pascals value to dB. Reference pressure defaults to 20 microPascals.
}
\examples{
PatodB(1,20)
93.9794
}
