% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseangle.R
\name{phaseangle}
\alias{phaseangle}
\title{Convert time difference to phase difference}
\usage{
phaseangle(tdiff = 0.001, freq = 100, unit = "deg")
}
\arguments{
\item{tdiff}{time difference in seconds}

\item{freq}{frequency in Hz}

\item{unit}{degrees ("deg") or radians ("rad")}

\item{f}{frequency}
}
\description{
Calculates the phase shift corresponding to a given time difference
}
