% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.R
\name{pressuredrop}
\alias{pressuredrop}
\title{Calculate pressure drop for spherical spreading. 6 dB per distance doubled}
\usage{
pressuredrop(r, Pa = NULL)
}
\arguments{
\item{r1}{distance}
}
\description{
Calculate pressure drop for spherical spreading. 6 dB per distance doubled
}
\examples{
pressuredrop(2)
6
}
