% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timedifference to phase shift.R
\name{timedifftophase}
\alias{timedifftophase}
\title{Convert time difference to phase difference}
\usage{
timedifftophase(tdiff = 0.001, freq = 100, unit = "deg")
}
\arguments{
\item{tdiff}{time difference in seconds}

\item{freq}{frequency in Hz}

\item{unit}{degrees or radians}

\item{f}{frequency}
}
\description{
Calculates the phase shift corresponding to a given time difference
}
