% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseshift.R
\name{timeshiftpressure}
\alias{timeshiftpressure}
\title{Phase shift}
\usage{
timeshiftpressure(Amplitude = 1, freq = 2, distance = 0, c = 340)
}
\arguments{
\item{c}{speed of sound}

\item{angle}{angle of sound in degrees. 0 is directly in front}

\item{d}{distance between two ears or receivers}

\item{f}{frequency}
}
\description{
Calculates the sine wave and time shifted time wave. Output is a list: S1 (first sound),
S2 (time shifted sound), Tadd (time added), Time (time)
}
\examples{
timeshiftpressure(Amplitude = 1, freq = 2, distance = 340, c = 340)
1.002374
}
