% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.R
\name{transmissionloss}
\alias{transmissionloss}
\title{Calculates transmission in 6 dB per distance doubled. Defaults to dB units, but Pa can be used}
\usage{
transmissionloss(distance = 2, initlevel = 100, unit = "dB")
}
\arguments{
\item{distance}{distance between source and receiver}

\item{initlevel}{sound level at source}

\item{unit}{units of dB or Pa}
}
\description{
Calculates transmission in 6 dB per distance doubled. Defaults to dB units, but Pa can be used
}
\examples{
pressuredrop(2)
6

pressuredrop(2, 100, unit = "Pa")
$Transmissionloss
[1] 0.5

$newlevel
[1] 50
}
