% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_misc_dopplershift.R
\name{dopplershift}
\alias{dopplershift}
\title{Phase shift}
\usage{
dopplershift(v1 = 10, v2 = 0, c = 340, freq = 300, direction = "towards")
}
\arguments{
\item{v1}{speed of first one}

\item{v2}{speed of second one}

\item{freq}{frequency}

\item{f}{frequency}
}
\description{
Calculates the sine wave and time shifted time wave. Output is a list: S1 (first sound),
S2 (time shifted sound), Tadd (time added), Time (time)
}
\examples{
timeshiftpressure(Amplitude = 1, freq = 2, distance = 340, c = 340)
1.002374
}
