% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human audiogram.R
\name{human_audiogram}
\alias{human_audiogram}
\title{Creates a dataframe with a human audiogram for reference from Jackson et al. 1999}
\usage{
human_audiogram()
}
\description{
Creates a dataframe with a human audiogram for reference from Jackson et al. 1999
}
